Backend songs album

See merge request nhatnk/amusic!52

+++ app/Http/Controllers/SongController.php
@@ -8,55 +8,58 @@ use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Storage;
 use App\Song;
 use App\User;
+use App\Album;
 
 class SongController extends Controller
 {
-    public function create(){
-
-        return view('songs.upload');
-
+    public function create(Request $request){
+        if(isset($request->id)){
+            $album=Album::find($request->id);
+        }
+        return view('songs.upload',compact('album'));
     }
 
-    /**
-     * @param Request $request
-     * @return \Illuminate\Http\RedirectResponse
-     */
     public function upload(Request $request){
         $this->validate($request,[
             'name' => 'required|min:3|max:50',
-            'audio' => 'required|mimes:mp3,mpga',
+            'audio' => 'required|mimes:mpga',
             'image' => 'mimes:jpeg,jpg,png,svg'
 
         ]);
 
         $song = new Song();
-
         $song->name = $request->input('name');
-
         $song->lyric = $request->input('lyric');
-
         $song->description = $request->input('description');
-
         if ($request->hasFile('image')){
             $song->image = $request->file('image')->store('image_songs/' . auth()->id(),'public');
         }
-
         if ($request->hasFile('audio')) {
             $song->audio = $request->file('audio')->store('audio_songs/' . auth()->id(),'public');
         }
-
         $song->user_id = Auth::id();
-
         $song->save();
 
-        Session::flash('announcement','Add Success');
+        if ($request->album_id){
+            $album = Album::find($request->input('album_id'));
+            $album->songs()->attach($song->id);
+            Session::flash('announcement','Upload Success!');
+            return redirect()->route('album.detail_album',['id'=>$album->id]);
+        }
 
-        return redirect()->back();
+
+        if ($request->input('check') != 1) {
+            Session::flash('announcement','Add Success');
+            return redirect()->route('song.list');
+        } else {
+            Session::flash('announcement','Add Success');
+            return redirect()->back();
+        }
     }
 
     public function index(){
 
-        $songs = Song::paginate(10);
+        $songs = Song::orderBy('id', 'desc')->paginate(10);
 
         if ($songs){
 
@@ -71,11 +74,15 @@ class SongController extends Controller
     }
 
     public function detailSong($id){
-
+ 
         $detail_song = Song::with('user')->find($id);
+
         if ($detail_song){
-            return view('songs.details_song', compact('detail_song','size_mb'));
+
+            return view('songs.details_song', compact('detail_song'));
+
         }else{
+            
             abort('404');
         }
 
@@ -84,7 +91,9 @@ class SongController extends Controller
     public function delete($id){
         $song = Song::find($id);
         if ($song){
-            Storage::delete('public/'.$song->image);
+            if(isset($song->image)){
+                Storage::delete('public/'.$song->image);
+            }
             Storage::delete('public/'.$song->audio);
             $song->delete();
             Session::flash('announcement','Delete Success');
@@ -125,6 +134,6 @@ class SongController extends Controller
     public function search(Request $request) {
         $key = $request->input('key');
         $songs = Song::where('name', 'like', '%'.$key.'%')->paginate(10);
-        return view('songs.search', compact('key','songs'));
+        return view('songs.search', compact('key','songs','albums'));
     }
 }
